[#629] Refactor reorderAnswersRepeatIndex as util & reuse in DataCleaningWebform
